Expose v8::Global
My understanding is that v8::Global is the new (2016?) and "better" way to have
control over a v8 value's lifetime. v8::Global is "better" because it's moveable
(whereas Persistent is copyable).

Now, for Zig/C, it really doesn't matter. The two are essentially the same. We
could keep using Persistent, but Global is the recommended class. I imagine it's
more likely to see updates / improvements. Since I'm reworking v8 integration a
little, this seems like a good time to do it.